chore(dashboard): Move dashboard functionality to separate module for logs
Address better code organization, move the dashboard-related code into
separate sub-module in the logs crate.
